values passed are doubles, not integers, and they have weight

+++ src/main/java/Axon.java
@@ -1,3 +1,3 @@
 public interface Axon {
-    void feed(int i);
+    void feed(double i);
 }
